Consolidating down to FULL_DUPLEX_STREAMED supported ext-proc server (#672)
