+++ config/charts/inferencepool/templates/epp-deployment.yaml
@@ -35,9 +35,6 @@ spec:
         - "9003"
         - -metricsPort
         - "9090"
-        env:
-        - name: USE_STREAMING
-          value: "true"
         ports:
         - name: grpc
           containerPort: 9002
