+++ pkg/epp/handlers/server.go
@@ -18,14 +18,23 @@ package handlers
 
 import (
 	"context"
+	"encoding/json"
 	"io"
+	"math/rand"
+	"strconv"
+	"strings"
 	"time"
 
+	configPb "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
 	extProcPb "github.com/envoyproxy/go-control-plane/envoy/service/ext_proc/v3"
 	envoyTypePb "github.com/envoyproxy/go-control-plane/envoy/type/v3"
+	"github.com/go-logr/logr"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/types/known/structpb"
 	"sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/gateway-api-inference-extension/api/v1alpha2"
+	backendmetrics "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/backend/metrics"
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/datastore"
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/metrics"
 	schedulingtypes "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/scheduling/types"
@@ -33,8 +42,8 @@ import (
 	logutil "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/util/logging"
 )
 
-func NewServer(scheduler Scheduler, destinationEndpointHintMetadataNamespace, destinationEndpointHintKey string, datastore datastore.Datastore) *Server {
-	return &Server{
+func NewStreamingServer(scheduler Scheduler, destinationEndpointHintMetadataNamespace, destinationEndpointHintKey string, datastore datastore.Datastore) *StreamingServer {
+	return &StreamingServer{
 		scheduler:                                scheduler,
 		destinationEndpointHintMetadataNamespace: destinationEndpointHintMetadataNamespace,
 		destinationEndpointHintKey:               destinationEndpointHintKey,
@@ -44,7 +53,7 @@ func NewServer(scheduler Scheduler, destinationEndpointHintMetadataNamespace, de
 
 // Server implements the Envoy external processing server.
 // https://www.envoyproxy.io/docs/envoy/latest/api-v3/service/ext_proc/v3/external_processor.proto
-type Server struct {
+type StreamingServer struct {
 	scheduler Scheduler
 	// The key of the header to specify the target pod address. This value needs to match Envoy
 	// configuration.
@@ -59,27 +68,75 @@ type Scheduler interface {
 	Schedule(ctx context.Context, b *schedulingtypes.LLMRequest) (targetPod schedulingtypes.Pod, err error)
 }
 
-func (s *Server) Process(srv extProcPb.ExternalProcessor_ProcessServer) error {
+// RequestContext stores context information during the life time of an HTTP request.
+type RequestContext struct {
+	TargetPod                 string
+	TargetEndpoint            string
+	Model                     string
+	ResolvedTargetModel       string
+	RequestReceivedTimestamp  time.Time
+	ResponseCompleteTimestamp time.Time
+	RequestSize               int
+	Usage                     Usage
+	ResponseSize              int
+	ResponseComplete          bool
+	ResponseStatusCode        string
+	RequestRunning            bool
+
+	RequestState         StreamRequestState
+	modelServerStreaming bool
+
+	reqHeaderResp  *extProcPb.ProcessingResponse
+	reqBodyResp    *extProcPb.ProcessingResponse
+	reqTrailerResp *extProcPb.ProcessingResponse
+
+	respHeaderResp  *extProcPb.ProcessingResponse
+	respBodyResp    *extProcPb.ProcessingResponse
+	respTrailerResp *extProcPb.ProcessingResponse
+}
+
+type StreamRequestState int
+
+const (
+	RequestReceived                  StreamRequestState = 0
+	HeaderRequestResponseComplete    StreamRequestState = 1
+	BodyRequestResponsesComplete     StreamRequestState = 2
+	TrailerRequestResponsesComplete  StreamRequestState = 3
+	ResponseRecieved                 StreamRequestState = 4
+	HeaderResponseResponseComplete   StreamRequestState = 5
+	BodyResponseResponsesComplete    StreamRequestState = 6
+	TrailerResponseResponsesComplete StreamRequestState = 7
+)
+
+func (s *StreamingServer) Process(srv extProcPb.ExternalProcessor_ProcessServer) error {
 	ctx := srv.Context()
 	logger := log.FromContext(ctx)
-	loggerVerbose := logger.V(logutil.VERBOSE)
-	loggerVerbose.Info("Processing")
+	loggerTrace := logger.V(logutil.TRACE)
+	loggerTrace.Info("Processing")
 
 	// Create request context to share states during life time of an HTTP request.
 	// See https://github.com/envoyproxy/envoy/issues/17540.
-	reqCtx := &RequestContext{}
+	reqCtx := &RequestContext{
+		RequestState: RequestReceived,
+	}
 
-	// Create variable for error handling as each request should only report once for
-	// error metric. This doesn't cover the error "Cannot receive stream request" because
-	// such error might happen even the response is processed.
+	var body []byte
+	var requestBody, responseBody map[string]interface{}
+
+	// Create error handling var as each request should only report once for
+	// error metrics. This doesn't cover the error "Cannot receive stream request" because
+	// such errors might happen even though response is processed.
 	var err error
-	defer func(error) {
+	defer func(error, *RequestContext) {
 		if reqCtx.ResponseStatusCode != "" {
 			metrics.RecordRequestErrCounter(reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.ResponseStatusCode)
 		} else if err != nil {
 			metrics.RecordRequestErrCounter(reqCtx.Model, reqCtx.ResolvedTargetModel, errutil.CanonicalCode(err))
 		}
-	}(err)
+		if reqCtx.RequestRunning {
+			metrics.DecRunningRequests(reqCtx.Model)
+		}
+	}(err, reqCtx)
 
 	for {
 		select {
@@ -95,70 +152,306 @@ func (s *Server) Process(srv extProcPb.ExternalProcessor_ProcessServer) error {
 		if recvErr != nil {
 			// This error occurs very frequently, though it doesn't seem to have any impact.
 			// TODO Figure out if we can remove this noise.
-			loggerVerbose.Error(err, "Cannot receive stream request")
+			logger.V(logutil.DEFAULT).Error(err, "Cannot receive stream request")
 			return status.Errorf(codes.Unknown, "cannot receive stream request: %v", err)
 		}
 
-		var resp *extProcPb.ProcessingResponse
 		switch v := req.Request.(type) {
 		case *extProcPb.ProcessingRequest_RequestHeaders:
-			reqCtx.RequestReceivedTimestamp = time.Now()
-			resp = HandleRequestHeaders(ctx, reqCtx, req)
-			loggerVerbose.Info("Request context after HandleRequestHeaders", "context", reqCtx)
+			err = s.HandleRequestHeaders(ctx, reqCtx, v)
 		case *extProcPb.ProcessingRequest_RequestBody:
-			resp, err = s.HandleRequestBody(ctx, reqCtx, req)
-			if err == nil {
-				metrics.RecordRequestCounter(reqCtx.Model, reqCtx.ResolvedTargetModel)
-				metrics.RecordRequestSizes(reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.RequestSize)
+			loggerTrace.Info("Incoming body chunk", "EoS", v.RequestBody.EndOfStream)
+			// In the stream case, we can receive multiple request bodies.
+			body = append(body, v.RequestBody.Body...)
+
+			// Message is buffered, we can read and decode.
+			if v.RequestBody.EndOfStream {
+				loggerTrace.Info("decoding")
+				err = json.Unmarshal(body, &requestBody)
+				if err != nil {
+					logger.V(logutil.DEFAULT).Error(err, "Error unmarshaling request body")
+				}
+
+				// Body stream complete. Allocate empty slice for response to use.
+				body = []byte{}
+
+				reqCtx, err = s.HandleRequestBody(ctx, reqCtx, req, requestBody)
+				if err != nil {
+					logger.V(logutil.DEFAULT).Error(err, "Error handling body")
+				} else {
+					metrics.RecordRequestCounter(reqCtx.Model, reqCtx.ResolvedTargetModel)
+					metrics.RecordRequestSizes(reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.RequestSize)
+				}
 			}
-			loggerVerbose.Info("Request context after HandleRequestBody", "context", reqCtx)
+		case *extProcPb.ProcessingRequest_RequestTrailers:
+			// This is currently unused.
 		case *extProcPb.ProcessingRequest_ResponseHeaders:
-			resp, err = s.HandleResponseHeaders(ctx, reqCtx, req)
-			loggerVerbose.Info("Request context after HandleResponseHeaders", "context", reqCtx)
+			for _, header := range v.ResponseHeaders.Headers.GetHeaders() {
+				value := string(header.RawValue)
+
+				loggerTrace.Info("header", "key", header.Key, "value", value)
+				if header.Key == "status" && value != "200" {
+					reqCtx.ResponseStatusCode = errutil.ModelServerError
+				} else if header.Key == "content-type" && strings.Contains(value, "text/event-stream") {
+					reqCtx.modelServerStreaming = true
+					loggerTrace.Info("model server is streaming response")
+				}
+			}
+			reqCtx.RequestState = ResponseRecieved
+			reqCtx.respHeaderResp = &extProcPb.ProcessingResponse{
+				Response: &extProcPb.ProcessingResponse_ResponseHeaders{
+					ResponseHeaders: &extProcPb.HeadersResponse{
+						Response: &extProcPb.CommonResponse{
+							HeaderMutation: &extProcPb.HeaderMutation{
+								SetHeaders: []*configPb.HeaderValueOption{
+									{
+										Header: &configPb.HeaderValue{
+											// This is for debugging purpose only.
+											Key:      "x-went-into-resp-headers",
+											RawValue: []byte("true"),
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			}
+
 		case *extProcPb.ProcessingRequest_ResponseBody:
-			// Don't send a 500 on a response error. Just let the message passthrough and log our error for debugging purposes.
-			// We assume the body is valid JSON, err messages are not guaranteed to be json, and so capturing and sending a 500 obfuscates the response message.
-			// using the standard 'err' var will send an immediate error response back to the caller.
-			var responseErr error
-			resp, responseErr = s.HandleResponseBody(ctx, reqCtx, req)
-			if responseErr != nil {
-				logger.V(logutil.DEFAULT).Error(responseErr, "Failed to process response body", "request", req)
-			} else if reqCtx.ResponseComplete {
-				reqCtx.ResponseCompleteTimestamp = time.Now()
-				metrics.RecordRequestLatencies(ctx, reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.RequestReceivedTimestamp, reqCtx.ResponseCompleteTimestamp)
-				metrics.RecordResponseSizes(reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.ResponseSize)
-				metrics.RecordInputTokens(reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.Usage.PromptTokens)
-				metrics.RecordOutputTokens(reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.Usage.CompletionTokens)
-				metrics.RecordNormalizedTimePerOutputToken(ctx, reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.RequestReceivedTimestamp, reqCtx.ResponseCompleteTimestamp, reqCtx.Usage.CompletionTokens)
-			}
 			if reqCtx.modelServerStreaming {
-				logger.V(logutil.DEBUG).Info("Request context after HandleResponseBody", "context", reqCtx)
+				// Currently we punt on response parsing if the modelServer is streaming, and we just passthrough.
+
+				responseText := string(v.ResponseBody.Body)
+				s.HandleResponseBodyModelStreaming(ctx, reqCtx, responseText)
+				if v.ResponseBody.EndOfStream {
+					loggerTrace.Info("stream completed")
+
+					reqCtx.ResponseCompleteTimestamp = time.Now()
+					metrics.RecordRequestLatencies(ctx, reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.RequestReceivedTimestamp, reqCtx.ResponseCompleteTimestamp)
+					metrics.RecordResponseSizes(reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.ResponseSize)
+				}
+
+				reqCtx.respBodyResp = &extProcPb.ProcessingResponse{
+					Response: &extProcPb.ProcessingResponse_ResponseBody{
+						ResponseBody: &extProcPb.BodyResponse{
+							Response: &extProcPb.CommonResponse{
+								BodyMutation: &extProcPb.BodyMutation{
+									Mutation: &extProcPb.BodyMutation_StreamedResponse{
+										StreamedResponse: &extProcPb.StreamedBodyResponse{
+											Body:        v.ResponseBody.Body,
+											EndOfStream: v.ResponseBody.EndOfStream,
+										},
+									},
+								},
+							},
+						},
+					},
+				}
 			} else {
-				loggerVerbose.Info("Request context after HandleResponseBody", "context", reqCtx)
+				body = append(body, v.ResponseBody.Body...)
+
+				// Message is buffered, we can read and decode.
+				if v.ResponseBody.EndOfStream {
+					loggerTrace.Info("stream completed")
+					// Don't send a 500 on a response error. Just let the message passthrough and log our error for debugging purposes.
+					// We assume the body is valid JSON, err messages are not guaranteed to be json, and so capturing and sending a 500 obfuscates the response message.
+					// using the standard 'err' var will send an immediate error response back to the caller.
+					var responseErr error
+					responseErr = json.Unmarshal(body, &responseBody)
+					if responseErr != nil {
+						logger.V(logutil.DEFAULT).Error(responseErr, "Error unmarshaling request body")
+					}
+
+					reqCtx, responseErr = s.HandleResponseBody(ctx, reqCtx, responseBody)
+					if responseErr != nil {
+						logger.V(logutil.DEFAULT).Error(responseErr, "Failed to process response body", "request", req)
+					} else if reqCtx.ResponseComplete {
+						reqCtx.ResponseCompleteTimestamp = time.Now()
+						metrics.RecordRequestLatencies(ctx, reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.RequestReceivedTimestamp, reqCtx.ResponseCompleteTimestamp)
+						metrics.RecordResponseSizes(reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.ResponseSize)
+						metrics.RecordInputTokens(reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.Usage.PromptTokens)
+						metrics.RecordOutputTokens(reqCtx.Model, reqCtx.ResolvedTargetModel, reqCtx.Usage.CompletionTokens)
+					}
+				}
 			}
-		default:
-			logger.V(logutil.DEFAULT).Error(nil, "Unknown Request type", "request", v)
-			return status.Error(codes.Unknown, "unknown request type")
+		case *extProcPb.ProcessingRequest_ResponseTrailers:
+			// This is currently unused.
 		}
 
+		// Handle the err and fire an immediate response.
 		if err != nil {
 			logger.V(logutil.DEFAULT).Error(err, "Failed to process request", "request", req)
-			resp, err = BuildErrResponse(err)
+			resp, err := BuildErrResponse(err)
 			if err != nil {
 				return err
 			}
+			if err := srv.Send(resp); err != nil {
+				logger.V(logutil.DEFAULT).Error(err, "Send failed")
+				return status.Errorf(codes.Unknown, "failed to send response back to Envoy: %v", err)
+			}
+			return nil
 		}
+		loggerTrace.Info("checking", "request state", reqCtx.RequestState)
+		if err := reqCtx.updateStateAndSendIfNeeded(srv, logger); err != nil {
+			return err
+		}
+	}
+}
 
-		if !reqCtx.modelServerStreaming {
-			loggerVerbose.Info("Response generated", "response", resp)
-		} else {
-			logger.V(logutil.DEBUG).Info("Response generated", "response", resp)
+// updateStateAndSendIfNeeded checks state and can send mutiple responses in a single pass, but only if ordered properly.
+// Order of requests matter in FULL_DUPLEX_STREAMING. For both request and response, the order of response sent back MUST be: Header->Body->Trailer, with trailer being optional.
+func (r *RequestContext) updateStateAndSendIfNeeded(srv extProcPb.ExternalProcessor_ProcessServer, logger logr.Logger) error {
+	loggerTrace := logger.V(logutil.TRACE)
+	// No switch statement as we could send multiple responses in one pass.
+	if r.RequestState == RequestReceived && r.reqHeaderResp != nil {
+		loggerTrace.Info("Sending request header response", "obj", r.reqHeaderResp)
+		if err := srv.Send(r.reqHeaderResp); err != nil {
+			logger.V(logutil.DEFAULT).Error(err, "error sending response")
+			return status.Errorf(codes.Unknown, "failed to send response back to Envoy: %v", err)
 		}
-		if err := srv.Send(resp); err != nil {
-			logger.V(logutil.DEFAULT).Error(err, "Send failed")
+		r.RequestState = HeaderRequestResponseComplete
+	}
+	if r.RequestState == HeaderRequestResponseComplete && r.reqBodyResp != nil {
+		loggerTrace.Info("Sending request body response")
+		if err := srv.Send(r.reqBodyResp); err != nil {
+			return status.Errorf(codes.Unknown, "failed to send response back to Envoy: %v", err)
+		}
+		r.RequestState = BodyRequestResponsesComplete
+		metrics.IncRunningRequests(r.Model)
+		r.RequestRunning = true
+		// Dump the response so a new stream message can begin
+		r.reqBodyResp = nil
+	}
+	if r.RequestState == BodyRequestResponsesComplete && r.reqTrailerResp != nil {
+		// Trailers in requests are not guaranteed
+		if err := srv.Send(r.reqHeaderResp); err != nil {
 			return status.Errorf(codes.Unknown, "failed to send response back to Envoy: %v", err)
 		}
 	}
+	if r.RequestState == ResponseRecieved && r.respHeaderResp != nil {
+		loggerTrace.Info("Sending response header response", "obj", r.respHeaderResp)
+		if err := srv.Send(r.respHeaderResp); err != nil {
+			return status.Errorf(codes.Unknown, "failed to send response back to Envoy: %v", err)
+		}
+		r.RequestState = HeaderResponseResponseComplete
+	}
+	if r.RequestState == HeaderResponseResponseComplete && r.respBodyResp != nil {
+		loggerTrace.Info("Sending response body response")
+		if err := srv.Send(r.respBodyResp); err != nil {
+			return status.Errorf(codes.Unknown, "failed to send response back to Envoy: %v", err)
+		}
+
+		body := r.respBodyResp.Response.(*extProcPb.ProcessingResponse_ResponseBody)
+		if body.ResponseBody.Response.GetBodyMutation().GetStreamedResponse().GetEndOfStream() {
+			r.RequestState = BodyResponseResponsesComplete
+		}
+		// Dump the response so a new stream message can begin
+		r.respBodyResp = nil
+	}
+	if r.RequestState == BodyResponseResponsesComplete && r.respTrailerResp != nil {
+		// Trailers in requests are not guaranteed
+		if err := srv.Send(r.reqHeaderResp); err != nil {
+			return status.Errorf(codes.Unknown, "failed to send response back to Envoy: %v", err)
+		}
+	}
+	return nil
+}
+
+func (s *StreamingServer) populateRequestHeaderResponse(reqCtx *RequestContext, endpoint string, requestBodyLength int) {
+	headers := []*configPb.HeaderValueOption{
+		{
+			Header: &configPb.HeaderValue{
+				Key:      s.destinationEndpointHintKey,
+				RawValue: []byte(endpoint),
+			},
+		},
+	}
+	if requestBodyLength > 0 {
+		// We need to update the content length header if the body is mutated, see Envoy doc:
+		// https://www.envoyproxy.io/docs/envoy/latest/api-v3/extensions/filters/http/ext_proc/v3/processing_mode.proto
+		headers = append(headers, &configPb.HeaderValueOption{
+			Header: &configPb.HeaderValue{
+				Key:      "Content-Length",
+				RawValue: []byte(strconv.Itoa(requestBodyLength)),
+			},
+		})
+	}
+
+	targetEndpointValue := &structpb.Struct{
+		Fields: map[string]*structpb.Value{
+			s.destinationEndpointHintKey: {
+				Kind: &structpb.Value_StringValue{
+					StringValue: endpoint,
+				},
+			},
+		},
+	}
+	dynamicMetadata := targetEndpointValue
+	if s.destinationEndpointHintMetadataNamespace != "" {
+		// If a namespace is defined, wrap the selected endpoint with that.
+		dynamicMetadata = &structpb.Struct{
+			Fields: map[string]*structpb.Value{
+				s.destinationEndpointHintMetadataNamespace: {
+					Kind: &structpb.Value_StructValue{
+						StructValue: targetEndpointValue,
+					},
+				},
+			},
+		}
+	}
+
+	reqCtx.reqHeaderResp = &extProcPb.ProcessingResponse{
+		Response: &extProcPb.ProcessingResponse_RequestHeaders{
+			RequestHeaders: &extProcPb.HeadersResponse{
+				Response: &extProcPb.CommonResponse{
+					ClearRouteCache: true,
+					HeaderMutation: &extProcPb.HeaderMutation{
+						SetHeaders: headers,
+					},
+				},
+			},
+		},
+		DynamicMetadata: dynamicMetadata,
+	}
+}
+
+func RandomWeightedDraw(logger logr.Logger, model *v1alpha2.InferenceModel, seed int64) string {
+	// TODO: after we are down to 1 server implementation, make these methods a part of the struct
+	// and handle random seeding on the struct.
+	source := rand.NewSource(rand.Int63())
+	if seed > 0 {
+		source = rand.NewSource(seed)
+	}
+	r := rand.New(source)
+
+	// all the weight values are nil, then we should return random model name
+	if model.Spec.TargetModels[0].Weight == nil {
+		index := r.Int31n(int32(len(model.Spec.TargetModels)))
+		return model.Spec.TargetModels[index].Name
+	}
+
+	var weights int32
+	for _, model := range model.Spec.TargetModels {
+		weights += *model.Weight
+	}
+	logger.V(logutil.TRACE).Info("Weights for model computed", "model", model.Name, "weights", weights)
+	randomVal := r.Int31n(weights)
+	// TODO: optimize this without using loop
+	for _, model := range model.Spec.TargetModels {
+		if randomVal < *model.Weight {
+			return model.Name
+		}
+		randomVal -= *model.Weight
+	}
+	return ""
+}
+
+func GetRandomPod(ds datastore.Datastore) *backendmetrics.Pod {
+	pods := ds.PodGetAll()
+	number := rand.Intn(len(pods))
+	pod := pods[number]
+	return pod.GetPod()
 }
 
 func BuildErrResponse(err error) (*extProcPb.ProcessingResponse, error) {
@@ -214,43 +507,3 @@ func BuildErrResponse(err error) (*extProcPb.ProcessingResponse, error) {
 	}
 	return resp, nil
 }
-
-// RequestContext stores context information during the life time of an HTTP request.
-type RequestContext struct {
-	TargetPod                 string
-	TargetEndpoint            string
-	Model                     string
-	ResolvedTargetModel       string
-	RequestReceivedTimestamp  time.Time
-	ResponseCompleteTimestamp time.Time
-	RequestSize               int
-	Usage                     Usage
-	ResponseSize              int
-	ResponseComplete          bool
-	ResponseStatusCode        string
-	RequestRunning            bool
-
-	RequestState         StreamRequestState
-	modelServerStreaming bool
-
-	reqHeaderResp  *extProcPb.ProcessingResponse
-	reqBodyResp    *extProcPb.ProcessingResponse
-	reqTrailerResp *extProcPb.ProcessingResponse
-
-	respHeaderResp  *extProcPb.ProcessingResponse
-	respBodyResp    *extProcPb.ProcessingResponse
-	respTrailerResp *extProcPb.ProcessingResponse
-}
-
-type StreamRequestState int
-
-const (
-	RequestReceived                  StreamRequestState = 0
-	HeaderRequestResponseComplete    StreamRequestState = 1
-	BodyRequestResponsesComplete     StreamRequestState = 2
-	TrailerRequestResponsesComplete  StreamRequestState = 3
-	ResponseRecieved                 StreamRequestState = 4
-	HeaderResponseResponseComplete   StreamRequestState = 5
-	BodyResponseResponsesComplete    StreamRequestState = 6
-	TrailerResponseResponsesComplete StreamRequestState = 7
-)
