+++ test/integration/epp/hermetic_test.go
@@ -92,325 +92,6 @@ func TestMain(m *testing.M) {
 	os.Exit(code)
 }
 
-func TestKubeInferenceModelRequest(t *testing.T) {
-	tests := []struct {
-		name              string
-		req               *extProcPb.ProcessingRequest
-		pods              map[backendmetrics.Pod]*backendmetrics.Metrics
-		wantHeaders       []*configPb.HeaderValueOption
-		wantMetadata      *structpb.Struct
-		wantBody          []byte
-		wantMetrics       string
-		wantErr           bool
-		immediateResponse *extProcPb.ImmediateResponse
-	}{
-		{
-			name: "select lower queue and kv cache, no active lora",
-			req:  integrationutils.GenerateRequest(logger, "test1", "my-model"),
-			// pod-1 will be picked because it has relatively low queue size and low KV cache.
-			pods: map[backendmetrics.Pod]*backendmetrics.Metrics{
-				fakePod(0): {
-					WaitingQueueSize:    3,
-					KVCacheUsagePercent: 0.2,
-				},
-				fakePod(1): {
-					WaitingQueueSize:    0,
-					KVCacheUsagePercent: 0.1,
-				},
-				fakePod(2): {
-					WaitingQueueSize:    10,
-					KVCacheUsagePercent: 0.2,
-				},
-			},
-			wantHeaders: []*configPb.HeaderValueOption{
-				{
-					Header: &configPb.HeaderValue{
-						Key:      runserver.DefaultDestinationEndpointHintKey,
-						RawValue: []byte("192.168.1.2:8000"),
-					},
-				},
-				{
-					Header: &configPb.HeaderValue{
-						Key:      "Content-Length",
-						RawValue: []byte("76"),
-					},
-				},
-			},
-			wantMetadata: makeMetadata("192.168.1.2:8000"),
-			wantBody:     []byte("{\"max_tokens\":100,\"model\":\"my-model-12345\",\"prompt\":\"test1\",\"temperature\":0}"),
-			wantMetrics: `
-			# HELP inference_model_request_total [ALPHA] Counter of inference model requests broken out for each model and target model.
-			# TYPE inference_model_request_total counter
-			inference_model_request_total{model_name="my-model",target_model_name="my-model-12345"} 1
-			`,
-			wantErr: false,
-		},
-		{
-			name: "select active lora, low queue",
-			req:  integrationutils.GenerateRequest(logger, "test2", "sql-lora"),
-			// pod-1 will be picked because it has relatively low queue size, with the requested
-			// model being active, and has low KV cache.
-			pods: map[backendmetrics.Pod]*backendmetrics.Metrics{
-				fakePod(0): {
-					WaitingQueueSize:    0,
-					KVCacheUsagePercent: 0.2,
-					ActiveModels: map[string]int{
-						"foo": 1,
-						"bar": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-				fakePod(1): {
-					WaitingQueueSize:    0,
-					KVCacheUsagePercent: 0.1,
-					ActiveModels: map[string]int{
-						"foo":            1,
-						"sql-lora-1fdg2": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-				fakePod(2): {
-					WaitingQueueSize:    10,
-					KVCacheUsagePercent: 0.2,
-					ActiveModels: map[string]int{
-						"foo": 1,
-						"bar": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-			},
-			wantHeaders: []*configPb.HeaderValueOption{
-				{
-					Header: &configPb.HeaderValue{
-						Key:      runserver.DefaultDestinationEndpointHintKey,
-						RawValue: []byte("192.168.1.2:8000"),
-					},
-				},
-				{
-					Header: &configPb.HeaderValue{
-						Key:      "Content-Length",
-						RawValue: []byte("76"),
-					},
-				},
-			},
-			wantMetadata: makeMetadata("192.168.1.2:8000"),
-			wantBody:     []byte("{\"max_tokens\":100,\"model\":\"sql-lora-1fdg2\",\"prompt\":\"test2\",\"temperature\":0}"),
-			wantMetrics: `
-			# HELP inference_model_request_total [ALPHA] Counter of inference model requests broken out for each model and target model.
-			# TYPE inference_model_request_total counter
-			inference_model_request_total{model_name="sql-lora",target_model_name="sql-lora-1fdg2"} 1
-			`,
-			wantErr: false,
-		},
-		{
-			name: "select no lora despite active model, avoid excessive queue size",
-			req:  integrationutils.GenerateRequest(logger, "test3", "sql-lora"),
-			// pod-2 will be picked despite it NOT having the requested model being active
-			// as it's above the affinity for queue size. Also is critical, so we should
-			// still honor request despite all queues > 5
-			pods: map[backendmetrics.Pod]*backendmetrics.Metrics{
-				fakePod(0): {
-					WaitingQueueSize:    10,
-					KVCacheUsagePercent: 0.2,
-					ActiveModels: map[string]int{
-						"foo": 1,
-						"bar": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-				fakePod(1): {
-					WaitingQueueSize:    200,
-					KVCacheUsagePercent: 0.1,
-					ActiveModels: map[string]int{
-						"foo":            1,
-						"sql-lora-1fdg2": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-				fakePod(2): {
-					WaitingQueueSize:    6,
-					KVCacheUsagePercent: 0.2,
-					ActiveModels: map[string]int{
-						"foo": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-			},
-			wantHeaders: []*configPb.HeaderValueOption{
-				{
-					Header: &configPb.HeaderValue{
-						Key:      runserver.DefaultDestinationEndpointHintKey,
-						RawValue: []byte("192.168.1.3:8000"),
-					},
-				},
-				{
-					Header: &configPb.HeaderValue{
-						Key:      "Content-Length",
-						RawValue: []byte("76"),
-					},
-				},
-			},
-			wantMetadata: makeMetadata("192.168.1.3:8000"),
-			wantBody:     []byte("{\"max_tokens\":100,\"model\":\"sql-lora-1fdg2\",\"prompt\":\"test3\",\"temperature\":0}"),
-			wantMetrics: `
-			# HELP inference_model_request_total [ALPHA] Counter of inference model requests broken out for each model and target model.
-			# TYPE inference_model_request_total counter
-			inference_model_request_total{model_name="sql-lora",target_model_name="sql-lora-1fdg2"} 1
-			`,
-			wantErr: false,
-		},
-		{
-			name: "noncritical and all models past threshold, shed request",
-			req:  integrationutils.GenerateRequest(logger, "test4", "sql-lora-sheddable"),
-			// no pods will be picked as all models are either above kv threshold,
-			// queue threshold, or both.
-			pods: map[backendmetrics.Pod]*backendmetrics.Metrics{
-				fakePod(0): {
-					WaitingQueueSize:    6,
-					KVCacheUsagePercent: 0.2,
-					ActiveModels: map[string]int{
-						"foo":            1,
-						"bar":            1,
-						"sql-lora-1fdg3": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-				fakePod(1): {
-					WaitingQueueSize:    0,
-					KVCacheUsagePercent: 0.85,
-					ActiveModels: map[string]int{
-						"foo":            1,
-						"sql-lora-1fdg3": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-				fakePod(2): {
-					WaitingQueueSize:    10,
-					KVCacheUsagePercent: 0.9,
-					ActiveModels: map[string]int{
-						"foo":            1,
-						"sql-lora-1fdg3": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-			},
-			wantHeaders:  []*configPb.HeaderValueOption{},
-			wantMetadata: &structpb.Struct{},
-			wantBody:     []byte(""),
-			wantErr:      false,
-			immediateResponse: &extProcPb.ImmediateResponse{
-				Status: &envoyTypePb.HttpStatus{
-					Code: envoyTypePb.StatusCode_TooManyRequests,
-				},
-			},
-			wantMetrics: "",
-		},
-		{
-			name: "noncritical, but one server has capacity, do not shed",
-			req:  integrationutils.GenerateRequest(logger, "test5", "sql-lora-sheddable"),
-			// pod 0 will be picked as all other models are above threshold
-			pods: map[backendmetrics.Pod]*backendmetrics.Metrics{
-				fakePod(0): {
-					WaitingQueueSize:    4,
-					KVCacheUsagePercent: 0.2,
-					ActiveModels: map[string]int{
-						"foo":            1,
-						"bar":            1,
-						"sql-lora-1fdg3": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-				fakePod(1): {
-					WaitingQueueSize:    0,
-					KVCacheUsagePercent: 0.85,
-					ActiveModels: map[string]int{
-						"foo":            1,
-						"sql-lora-1fdg3": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-				fakePod(2): {
-					WaitingQueueSize:    10,
-					KVCacheUsagePercent: 0.9,
-					ActiveModels: map[string]int{
-						"foo":            1,
-						"sql-lora-1fdg3": 1,
-					},
-					WaitingModels: map[string]int{},
-				},
-			},
-			wantHeaders: []*configPb.HeaderValueOption{
-				{
-					Header: &configPb.HeaderValue{
-						Key:      runserver.DefaultDestinationEndpointHintKey,
-						RawValue: []byte("192.168.1.1:8000"),
-					},
-				},
-				{
-					Header: &configPb.HeaderValue{
-						Key:      "Content-Length",
-						RawValue: []byte("76"),
-					},
-				},
-			},
-			wantMetadata: makeMetadata("192.168.1.1:8000"),
-			wantBody:     []byte("{\"max_tokens\":100,\"model\":\"sql-lora-1fdg3\",\"prompt\":\"test5\",\"temperature\":0}"),
-			wantMetrics: `
-			# HELP inference_model_request_total [ALPHA] Counter of inference model requests broken out for each model and target model.
-			# TYPE inference_model_request_total counter
-			inference_model_request_total{model_name="sql-lora-sheddable",target_model_name="sql-lora-1fdg3"} 1
-			`,
-			wantErr: false,
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			client, cleanup := setUpHermeticServer(t, test.pods, false)
-			t.Cleanup(cleanup)
-			want := &extProcPb.ProcessingResponse{
-				Response: &extProcPb.ProcessingResponse_RequestBody{
-					RequestBody: &extProcPb.BodyResponse{
-						Response: &extProcPb.CommonResponse{
-							HeaderMutation: &extProcPb.HeaderMutation{
-								SetHeaders: test.wantHeaders,
-							},
-							BodyMutation: &extProcPb.BodyMutation{
-								Mutation: &extProcPb.BodyMutation_Body{
-									Body: test.wantBody,
-								},
-							},
-						},
-					},
-				},
-				DynamicMetadata: test.wantMetadata,
-			}
-			res, err := integrationutils.SendRequest(t, client, test.req)
-
-			if err != nil && !test.wantErr {
-				t.Errorf("Unexpected error, got: %v, want error: %v", err, test.wantErr)
-			}
-			if test.immediateResponse != nil {
-				want = &extProcPb.ProcessingResponse{
-					Response: &extProcPb.ProcessingResponse_ImmediateResponse{
-						ImmediateResponse: test.immediateResponse,
-					},
-				}
-			}
-			if diff := cmp.Diff(want, res, protocmp.Transform()); diff != "" {
-				t.Errorf("Unexpected response, (-want +got): %v", diff)
-			}
-
-			if test.wantMetrics != "" {
-				if err := metricsutils.GatherAndCompare(legacyregistry.DefaultGatherer, strings.NewReader(test.wantMetrics), "inference_model_request_total"); err != nil {
-					t.Error(err)
-				}
-			}
-
-			legacyregistry.Reset()
-		})
-	}
-}
-
 func TestFullDuplexStreamed_KubeInferenceModelRequest(t *testing.T) {
 	tests := []struct {
 		name              string
